Finish
Implemented notification

+++ app/Jobs/DeletePendingViolations.php
@@ -2,7 +2,9 @@
 
 namespace App\Jobs;
 
+use App\Http\Resources\PendingViolationResource;
 use App\Models\EmployeeViolation;
+use App\Models\Log;
 use App\Models\PendingViolation;
 use App\Notifications\ViolationSuccess;
 use Illuminate\Bus\Queueable;
@@ -11,7 +13,7 @@ use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
-use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Request;
 
 class DeletePendingViolations implements ShouldQueue
 {
@@ -34,12 +36,24 @@ class DeletePendingViolations implements ShouldQueue
     public function handle(): void
     {
         $pendingViolations = PendingViolation::whereIn('employee_id', $this->pendingViolations)->get();
+        info('empty: ' . empty($pendingViolations) . ' count: ' . count($pendingViolations));
+        if (count($pendingViolations) === 0)  return;
+
+
+
         foreach ($pendingViolations as $data) {
             EmployeeViolation::create([
                 'employee_id' => $data->employee_id,
                 'violation_id' => $data->violation_id
             ]);
         }
+
+        Request::merge(['includeViolationType' => true]);
+
+        Log::create([
+            'type' => 'pendingViolation',
+            'data' => json_encode(PendingViolationResource::collection($pendingViolations))
+        ]);
 
         PendingViolation::whereIn('employee_id', $this->pendingViolations)->delete();
